+++ .idea/cmoneyProject_w2_citymanage.iml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" inherit-compiler-output="true">
+    <exclude-output />
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/cmoneyw1citymanage/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/cmoneyw1citymanage/src/main/java" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/cmoneyw2citymanage" isTestSource="false" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+  </component>
+</module>
