+++ cmoneyw1citymanage/src/main/java/cmoneyweek1citymanage/LumberCamp.java
@@ -1,4 +1,4 @@
-/*
+package cmoneyweek1citymanage;/*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
@@ -8,26 +8,22 @@
  *
  * @author kisso
  */
-public class House extends Building{
-    /**
-     * 升級所需時間
-     * 村民生命值
-     * 村民生產速度
-     */
-    private int villagerGenRate;
-    private int villagerLife;
+public class LumberCamp extends Building{
+    private final int upgradeResetTime = 30;
+    private int woodRate ;
 
-    public House(){
+    public LumberCamp(){
+
         //建築編號
-        this.setNumber(1);
+        this.setNumber(4);
         //建築名稱
-        this.setName("房屋");
+        this.setName("伐木場");
         //建築等級
         this.setBuildingLevel(1);
         //建築生命值
         this.setLife(10);
         //建造所需資源
-        this.setBuildResource(new Resource(10,0,0));
+        this.setBuildResource(new Resource(15,0,0));
         //建築狀態
         this.setBuildCheck(Building.BuildCheck.BUILDABLE);
         //建造所需時間
@@ -46,17 +42,17 @@ public class House extends Building{
         this.setUpgradeResetTime(30);
         //建築物功能開關
         this.setOnOff(false);
-        //自動生產消耗資源
-        this.setEffectResource(new Resource(1,0,0));
         //升級需要文明等級
         this.setUpNeedCivilLevel(2);
     }
 
-    public int getVillagerGenRate() {
-        villagerGenRate = this.getBuildingLevel()*2-1;
-        return villagerGenRate;
+    public int getWoodRate() {
+        
+        woodRate = this.getBuildingLevel()+1;
+        if(this.getBuildCheck().equals(Building.BuildCheck.UNBUILDABLE)){
+            woodRate ++;
+        }
+        return woodRate;
     }
-    
-    
 
 }
