+++ cmoneyw1citymanage/src/main/java/cmoneyweek1citymanage/MiningCamp.java
@@ -1,4 +1,4 @@
-/*
+package cmoneyweek1citymanage;/*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
