+++ cmoneyw1citymanage/src/main/java/cmoneyweek1citymanage/University.java
@@ -1,4 +1,4 @@
-/*
+package cmoneyweek1citymanage;/*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
@@ -8,48 +8,40 @@
  *
  * @author kisso
  */
-public class GasCamp extends Building{
-    private int gasRate ;
-
-    public GasCamp(){
+public class University extends Building{
+    public University(){
 
         //建築編號
-        this.setNumber(7);
+        this.setNumber(2);
         //建築名稱
-        this.setName("瓦斯廠");
+        this.setName("研究所");
         //建築等級
         this.setBuildingLevel(1);
         //建築生命值
-        this.setLife(20);
+        this.setLife(30);
         //建造所需資源
-        this.setBuildResource(new Resource(15,5,0));
+        this.setBuildResource(new Resource(10,5,0));
         //建築狀態
         this.setBuildCheck(Building.BuildCheck.BUILDABLE);
         //建造所需時間
-        this.setBuildNeedTime(1);
+        this.setBuildNeedTime(3);
         //建築建好時間
         this.setBuildTime(-1);
         //建造需要文明等級
-        this.setNeedCivilLevel(2);
+        this.setNeedCivilLevel(1);
         //升級所需資源
-        this.setUpgradeResource(new Resource(40,20,0));
+        this.setUpgradeResource(new Resource(50,20,0));
         //升級狀態
         this.setUpgradeCheck(Building.UpgradeCheck.UPGRADEABLE);
         //升級剩餘時間
-        this.setUpgradeNeedTime(30);
+        this.setUpgradeNeedTime(24);
         //升級所需時間
-        this.setUpgradeResetTime(30);
+        this.setUpgradeResetTime(24);
         //建築物功能開關
         this.setOnOff(false);
         //升級需要文明等級
-        this.setUpNeedCivilLevel(3);
+        this.setUpNeedCivilLevel(1);
     }
     
-    public int getGasRate() {
-        
-        gasRate = this.getBuildingLevel()+4;
-
-        return gasRate;
-    }
 
 }
